chore: split the menu config test by behaviour
- What: `tests/http/menu/config.test.ts` 322 → **113**, with gist fetching, retry
  and the source/reload endpoints moving to `config-gist-fetch.test.ts` (219).
  Allow-list entry removed, not lowered.

- Why: #2833, and CLAUDE.md's "nested, one behavior per file". This one needed a
  real split rather than a fixture extraction — four describe blocks covering two
  genuinely different areas, and pulling out the ~40 lines of setup alone would
  still have left it over 250.

- Impact: none. Same 125 tests pass in `tests/http/menu/` as before the split.

Dead imports pruned from both halves rather than carried along — each file now
imports only what it uses, verified symbol by symbol. A split that leaves the
original import block intact just moves the mess.

The gist half keeps the `afterEach` fetch restore, and says why in its header:
`--isolate` gives each FILE a fresh global, not each test, so a suite that
replaces `globalThis.fetch` still has to put it back between tests.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate tests/http/menu/
tests/build/` 141 pass.

Allow-list: 22 at #2833's filing → **15**.

Refs #2833
